+++ bethink/src/screens/reminders/RemindersScreen.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+export const RemindersScreen = () => {
+  return (
+    <View>
+      <Text>Reminders</Text>
+    </View>
+  );
+};
